Clean up primary-only snapshot in StorageManagerImpl scavenger instead
